move text line serialization tests to kernel tests
