+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -165,24 +165,6 @@ tf_py_test(
 )
 
 tf_py_test(
-    name = "textline_dataset_serialization_test",
-    size = "medium",
-    srcs = ["textline_dataset_serialization_test.py"],
-    shard_count = 4,
-    tags = [
-        "no_oss",
-        "no_pip",
-        "no_windows",
-    ],
-    deps = [
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python/data/experimental/kernel_tests:reader_dataset_ops_test_base",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-        "//tensorflow/python/data/ops:readers",
-    ],
-)
-
-tf_py_test(
     name = "tf_record_dataset_serialization_test",
     size = "medium",
     srcs = ["tf_record_dataset_serialization_test.py"],
