+++ tensorflow/python/data/kernel_tests/BUILD
@@ -758,13 +758,16 @@ tf_py_test(
     name = "text_line_dataset_test",
     size = "small",
     srcs = ["text_line_dataset_test.py"],
+    shard_count = 4,
     deps = [
+        ":checkpoint_test",
         ":test_base",
         "//tensorflow/python:array_ops",
         "//tensorflow/python:client_testlib",
         "//tensorflow/python:dtypes",
         "//tensorflow/python:errors",
         "//tensorflow/python:util",
+        "//tensorflow/python/data/experimental/kernel_tests:reader_dataset_ops_test_base",
         "//tensorflow/python/data/ops:iterator_ops",
         "//tensorflow/python/data/ops:readers",
         "//tensorflow/python/eager:context",
